backend/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
pydantic==2.6.4
python-dotenv==1.0.1
lxml==5.1.0
firebase-admin==6.5.0
